requirements.txt
accelerate==0.26.0
beautifulsoup4==4.9.3
bitsandbytes==0.41.3
datasets==2.18.0
deepface==0.0.91
evaluate==0.4.1
face_recognition==1.3.0
face_recognition_models==0.3.0
faiss-cpu
fasttext
google-api-core
google-api-python-client
google-cloud-vision
haversine
imagehash==4.3.1
Levenshtein==0.23.0
matplotlib
pandas
pillow==10.3.0
rouge==1.0.1
scikit-learn
scipy
sentence-transformers
sentencepiece
spacy==3.5.0
torch
tqdm
trafilatura
transformers
vllm
